FAN Model complete

+++ model/data/augment.py
@@ -327,4 +327,4 @@ class ToTensor:
         # numpy image: H x W x C
         # torch image: C X H X W
         image = image.transpose((2, 0, 1))
-        return torch.from_numpy(image).float(), torch.from_numpy(mask)
+        return torch.from_numpy(image).float(), torch.from_numpy(mask).float()
